--- a/cmd/lotus-shed/main.go
+++ b/cmd/lotus-shed/main.go
@@ -91,6 +91,7 @@ func main() {
 		FevmAnalyticsCmd,
 		mismatchesCmd,
 		blockCmd,
+		adlCmd,
 	}
 
 	app := &cli.App{