## Related Issues
<!-- Link issues that this PR might resolve/fix. If an issue doesn't exist, include a brief motivation for the change being made -->

## Proposed Changes
<!-- A clear list of the changes being made -->

## Additional Info
Current output
```
REMINDER: If you are running this, you likely want to ALSO run the continuity testing tool!
manifest cid: bafy2bzacecpuup2nsfmmrrkm4fb4bvm4gcmhvoslhzmnz36w5teecs2mrvhjg
2024-01-25T15:48:24.150+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Started 7 workers
2024-01-25T15:48:24.151+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Result writer started
2024-01-25T15:48:24.151+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Creating migration jobs
2024-01-25T15:48:34.156+0100    INFO    fil-consensus   filcns/upgrades.go:2418 98152 jobs created, 97145 done, 1007 pending after 10.005548708s (9709/s)
2024-01-25T15:48:44.156+0100    INFO    fil-consensus   filcns/upgrades.go:2418 185317 jobs created, 184310 done, 1007 pending after 20.00564325s (9213/s)
2024-01-25T15:48:54.157+0100    INFO    fil-consensus   filcns/upgrades.go:2418 301311 jobs created, 300304 done, 1007 pending after 30.006194333s (10008/s)
2024-01-25T15:49:04.157+0100    INFO    fil-consensus   filcns/upgrades.go:2418 386255 jobs created, 385248 done, 1007 pending after 40.006319167s (9630/s)
2024-01-25T15:49:14.157+0100    INFO    fil-consensus   filcns/upgrades.go:2418 469984 jobs created, 468977 done, 1007 pending after 50.006376833s (9378/s)
2024-01-25T15:49:24.162+0100    INFO    fil-consensus   filcns/upgrades.go:2418 570946 jobs created, 569939 done, 1007 pending after 1m0.011605083s (9497/s)
2024-01-25T15:49:34.162+0100    INFO    fil-consensus   filcns/upgrades.go:2418 637920 jobs created, 636913 done, 1007 pending after 1m10.011709875s (9097/s)
2024-01-25T15:49:44.162+0100    INFO    fil-consensus   filcns/upgrades.go:2418 728798 jobs created, 727791 done, 1007 pending after 1m20.011806708s (9096/s)
2024-01-25T15:49:54.162+0100    INFO    fil-consensus   filcns/upgrades.go:2418 791869 jobs created, 790862 done, 1007 pending after 1m30.011914208s (8786/s)
2024-01-25T15:50:04.163+0100    INFO    fil-consensus   filcns/upgrades.go:2418 882034 jobs created, 881027 done, 1007 pending after 1m40.012076125s (8809/s)
2024-01-25T15:50:14.163+0100    INFO    fil-consensus   filcns/upgrades.go:2418 956913 jobs created, 955906 done, 1007 pending after 1m50.012562875s (8689/s)
2024-01-25T15:50:24.163+0100    INFO    fil-consensus   filcns/upgrades.go:2418 1033197 jobs created, 1032190 done, 1007 pending after 2m0.012627292s (8601/s)
2024-01-25T15:50:34.163+0100    INFO    fil-consensus   filcns/upgrades.go:2418 1110621 jobs created, 1109614 done, 1007 pending after 2m10.012688625s (8535/s)
2024-01-25T15:50:44.163+0100    INFO    fil-consensus   filcns/upgrades.go:2418 1191577 jobs created, 1190570 done, 1007 pending after 2m20.01278225s (8503/s)
2024-01-25T15:50:54.163+0100    INFO    fil-consensus   filcns/upgrades.go:2418 1255255 jobs created, 1254248 done, 1007 pending after 2m30.01295775s (8361/s)
2024-01-25T15:51:04.164+0100    INFO    fil-consensus   filcns/upgrades.go:2418 1350326 jobs created, 1349319 done, 1007 pending after 2m40.013095875s (8433/s)
2024-01-25T15:51:14.164+0100    INFO    fil-consensus   filcns/upgrades.go:2418 1446653 jobs created, 1445646 done, 1007 pending after 2m50.0132085s (8503/s)
2024-01-25T15:51:24.164+0100    INFO    fil-consensus   filcns/upgrades.go:2418 1536712 jobs created, 1535705 done, 1007 pending after 3m0.013275417s (8531/s)
2024-01-25T15:51:34.164+0100    INFO    fil-consensus   filcns/upgrades.go:2418 1632378 jobs created, 1631371 done, 1007 pending after 3m10.013350958s (8586/s)
2024-01-25T15:51:44.164+0100    INFO    fil-consensus   filcns/upgrades.go:2418 1705779 jobs created, 1704772 done, 1007 pending after 3m20.0134205s (8523/s)
2024-01-25T15:51:54.164+0100    INFO    fil-consensus   filcns/upgrades.go:2418 1797971 jobs created, 1796964 done, 1007 pending after 3m30.013485208s (8556/s)
2024-01-25T15:52:04.164+0100    INFO    fil-consensus   filcns/upgrades.go:2418 1868942 jobs created, 1867935 done, 1007 pending after 3m40.013576375s (8490/s)
2024-01-25T15:52:14.165+0100    INFO    fil-consensus   filcns/upgrades.go:2418 1935723 jobs created, 1934716 done, 1007 pending after 3m50.014086708s (8411/s)
2024-01-25T15:52:24.165+0100    INFO    fil-consensus   filcns/upgrades.go:2418 2006573 jobs created, 2005566 done, 1007 pending after 4m0.014157708s (8356/s)
2024-01-25T15:52:34.165+0100    INFO    fil-consensus   filcns/upgrades.go:2418 2068037 jobs created, 2067030 done, 1007 pending after 4m10.014233208s (8268/s)
2024-01-25T15:52:44.165+0100    INFO    fil-consensus   filcns/upgrades.go:2418 2149619 jobs created, 2148707 done, 912 pending after 4m20.014354417s (8264/s)
2024-01-25T15:52:54.165+0100    INFO    fil-consensus   filcns/upgrades.go:2418 2199829 jobs created, 2198855 done, 974 pending after 4m30.0144465s (8143/s)
2024-01-25T15:53:04.167+0100    INFO    fil-consensus   filcns/upgrades.go:2418 2264154 jobs created, 2263147 done, 1007 pending after 4m40.01603275s (8082/s)
2024-01-25T15:53:14.167+0100    INFO    fil-consensus   filcns/upgrades.go:2418 2314791 jobs created, 2313784 done, 1007 pending after 4m50.0161565s (7978/s)
2024-01-25T15:53:24.168+0100    INFO    fil-consensus   filcns/upgrades.go:2418 2402823 jobs created, 2401816 done, 1007 pending after 5m0.016249333s (8006/s)
2024-01-25T15:53:34.169+0100    INFO    fil-consensus   filcns/upgrades.go:2418 2475169 jobs created, 2474162 done, 1007 pending after 5m10.018313125s (7981/s)
2024-01-25T15:53:44.169+0100    INFO    fil-consensus   filcns/upgrades.go:2418 2565681 jobs created, 2564674 done, 1007 pending after 5m20.018389083s (8014/s)
2024-01-25T15:53:54.169+0100    INFO    fil-consensus   filcns/upgrades.go:2418 2642662 jobs created, 2641655 done, 1007 pending after 5m30.018498542s (8005/s)
2024-01-25T15:54:04.169+0100    INFO    fil-consensus   filcns/upgrades.go:2418 2723608 jobs created, 2722601 done, 1007 pending after 5m40.018577s (8007/s)
2024-01-25T15:54:14.199+0100    INFO    fil-consensus   filcns/upgrades.go:2418 2778931 jobs created, 2777924 done, 1007 pending after 5m50.037615875s (7936/s)
2024-01-25T15:54:24.212+0100    INFO    fil-consensus   filcns/upgrades.go:2418 2849771 jobs created, 2848764 done, 1007 pending after 6m0.037717208s (7912/s)
2024-01-25T15:54:34.219+0100    INFO    fil-consensus   filcns/upgrades.go:2418 2936106 jobs created, 2935181 done, 925 pending after 6m10.037786458s (7932/s)
2024-01-25T15:54:35.165+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Done creating 2947604 migration jobs after 6m10.982977542s
2024-01-25T15:54:35.184+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Worker 1 done
2024-01-25T15:54:35.223+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Worker 4 done
2024-01-25T15:54:35.237+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Worker 6 done
2024-01-25T15:54:35.261+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Worker 0 done
2024-01-25T15:54:35.329+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Worker 5 done
2024-01-25T15:54:36.174+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Worker 2 done
2024-01-25T15:54:36.532+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Worker 3 done
2024-01-25T15:54:36.532+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Result writer wrote 2947603 results to state tree after 6m12.349480875s
2024-01-25T15:54:36.532+0100    INFO    fil-consensus   filcns/upgrades.go:2418 All workers done after 6m12.349460333s
2024-01-25T15:54:36.532+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Running deferred migrations
2024-01-25T15:54:36.532+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Running deferred migration for f05
2024-01-25T15:57:01.796+0100    INFO    fil-consensus   filcns/upgrades.go:2418 All 2947604 done after 8m37.605489417s (5695/s)
migration height  3596225
old cid  bafy2bzaceamqaqlxzhfvs2hf32gvtuqhhfvdbwh7ztrk36foixkimsthdf462
new cid  bafy2bzacedbolm3bvanxqy7b2n5xxovg2eel3ii35qlfunf2iktkteucd5nb2
completed round actual (without cache), took  8m39.557524208s
manifest cid: bafy2bzacecpuup2nsfmmrrkm4fb4bvm4gcmhvoslhzmnz36w5teecs2mrvhjg
2024-01-25T15:57:03.981+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Started 5 workers
2024-01-25T15:57:03.981+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Creating migration jobs
2024-01-25T15:57:03.982+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Result writer started
2024-01-25T16:02:03.985+0100    INFO    fil-consensus   filcns/upgrades.go:2418 1954661 jobs created, 1954656 done, 5 pending after 5m0.000420667s (6516/s)
2024-01-25T16:04:57.641+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Done creating 2947309 migration jobs after 7m53.655064875s
2024-01-25T16:04:57.642+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Worker 0 done
2024-01-25T16:04:57.642+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Worker 1 done
2024-01-25T16:04:57.743+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Worker 2 done
2024-01-25T16:04:58.312+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Worker 4 done
2024-01-25T16:04:58.516+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Worker 3 done
2024-01-25T16:04:58.516+0100    INFO    fil-consensus   filcns/upgrades.go:2418 All workers done after 7m54.529966834s
2024-01-25T16:04:58.516+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Result writer wrote 2947308 results to state tree after 7m54.52999675s
2024-01-25T16:04:58.516+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Running deferred migrations
2024-01-25T16:04:58.516+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Running deferred migration for f05
2024-01-25T16:07:27.372+0100    INFO    fil-consensus   filcns/upgrades.go:2418 All 2947309 done after 10m23.384064875s (4728/s)
completed premigration, took  10m25.223251917s
manifest cid: bafy2bzacecpuup2nsfmmrrkm4fb4bvm4gcmhvoslhzmnz36w5teecs2mrvhjg
2024-01-25T16:07:28.955+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Started 7 workers
2024-01-25T16:07:28.955+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Creating migration jobs
2024-01-25T16:07:28.955+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Result writer started
2024-01-25T16:07:38.955+0100    INFO    fil-consensus   filcns/upgrades.go:2418 98152 jobs created, 97145 done, 1007 pending after 10.00006925s (9714/s)
2024-01-25T16:07:48.955+0100    INFO    fil-consensus   filcns/upgrades.go:2418 187297 jobs created, 186290 done, 1007 pending after 20.000144875s (9314/s)
2024-01-25T16:07:58.957+0100    INFO    fil-consensus   filcns/upgrades.go:2418 316593 jobs created, 315586 done, 1007 pending after 30.001249s (10519/s)
2024-01-25T16:08:08.959+0100    INFO    fil-consensus   filcns/upgrades.go:2418 412824 jobs created, 411817 done, 1007 pending after 40.003168166s (10295/s)
2024-01-25T16:08:18.959+0100    INFO    fil-consensus   filcns/upgrades.go:2418 513133 jobs created, 512126 done, 1007 pending after 50.003259916s (10242/s)
2024-01-25T16:08:28.959+0100    INFO    fil-consensus   filcns/upgrades.go:2418 638799 jobs created, 638162 done, 637 pending after 1m0.003432541s (10635/s)
2024-01-25T16:08:38.959+0100    INFO    fil-consensus   filcns/upgrades.go:2418 768513 jobs created, 767506 done, 1007 pending after 1m10.003502875s (10964/s)
2024-01-25T16:08:48.959+0100    INFO    fil-consensus   filcns/upgrades.go:2418 856089 jobs created, 855082 done, 1007 pending after 1m20.003574541s (10688/s)
2024-01-25T16:08:58.964+0100    INFO    fil-consensus   filcns/upgrades.go:2418 960945 jobs created, 959938 done, 1007 pending after 1m30.00758175s (10665/s)
2024-01-25T16:09:08.966+0100    INFO    fil-consensus   filcns/upgrades.go:2418 1073829 jobs created, 1072822 done, 1007 pending after 1m40.009580375s (10727/s)
2024-01-25T16:09:18.966+0100    INFO    fil-consensus   filcns/upgrades.go:2418 1178621 jobs created, 1177801 done, 820 pending after 1m50.0097315s (10706/s)
2024-01-25T16:09:28.966+0100    INFO    fil-consensus   filcns/upgrades.go:2418 1241808 jobs created, 1240801 done, 1007 pending after 2m0.009873s (10339/s)
2024-01-25T16:09:38.966+0100    INFO    fil-consensus   filcns/upgrades.go:2418 1342711 jobs created, 1341704 done, 1007 pending after 2m10.010000666s (10320/s)
2024-01-25T16:09:48.967+0100    INFO    fil-consensus   filcns/upgrades.go:2418 1463172 jobs created, 1462165 done, 1007 pending after 2m20.010071208s (10443/s)
2024-01-25T16:09:58.967+0100    INFO    fil-consensus   filcns/upgrades.go:2418 1582286 jobs created, 1581279 done, 1007 pending after 2m30.010192s (10541/s)
2024-01-25T16:10:08.967+0100    INFO    fil-consensus   filcns/upgrades.go:2418 1687403 jobs created, 1687330 done, 73 pending after 2m40.01027375s (10545/s)
2024-01-25T16:10:18.967+0100    INFO    fil-consensus   filcns/upgrades.go:2418 1828209 jobs created, 1827202 done, 1007 pending after 2m50.010353208s (10748/s)
2024-01-25T16:10:28.968+0100    INFO    fil-consensus   filcns/upgrades.go:2418 1934049 jobs created, 1933042 done, 1007 pending after 3m0.01066725s (10738/s)
2024-01-25T16:10:38.969+0100    INFO    fil-consensus   filcns/upgrades.go:2418 2052296 jobs created, 2051289 done, 1007 pending after 3m10.012419875s (10796/s)
2024-01-25T16:10:48.970+0100    INFO    fil-consensus   filcns/upgrades.go:2418 2171952 jobs created, 2170945 done, 1007 pending after 3m20.012530958s (10854/s)
2024-01-25T16:10:58.970+0100    INFO    fil-consensus   filcns/upgrades.go:2418 2260316 jobs created, 2259309 done, 1007 pending after 3m30.012603625s (10758/s)
2024-01-25T16:11:08.971+0100    INFO    fil-consensus   filcns/upgrades.go:2418 2338685 jobs created, 2338480 done, 205 pending after 3m40.013232708s (10629/s)
2024-01-25T16:11:18.971+0100    INFO    fil-consensus   filcns/upgrades.go:2418 2463665 jobs created, 2462748 done, 917 pending after 3m50.013339625s (10707/s)
2024-01-25T16:11:28.974+0100    INFO    fil-consensus   filcns/upgrades.go:2418 2598835 jobs created, 2597828 done, 1007 pending after 4m0.016053958s (10824/s)
2024-01-25T16:11:38.976+0100    INFO    fil-consensus   filcns/upgrades.go:2418 2729130 jobs created, 2728123 done, 1007 pending after 4m10.018020625s (10912/s)
2024-01-25T16:11:48.976+0100    INFO    fil-consensus   filcns/upgrades.go:2418 2803049 jobs created, 2802042 done, 1007 pending after 4m20.018111s (10776/s)
2024-01-25T16:11:58.976+0100    INFO    fil-consensus   filcns/upgrades.go:2418 2904374 jobs created, 2903367 done, 1007 pending after 4m30.0182235s (10752/s)
2024-01-25T16:12:02.121+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Done creating 2947604 migration jobs after 4m33.163017958s
2024-01-25T16:12:02.121+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Worker 0 done
2024-01-25T16:12:02.469+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Worker 3 done
2024-01-25T16:12:02.480+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Worker 6 done
2024-01-25T16:12:02.618+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Worker 2 done
2024-01-25T16:12:02.960+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Worker 5 done
2024-01-25T16:12:03.072+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Worker 4 done
2024-01-25T16:12:03.287+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Worker 1 done
2024-01-25T16:12:03.287+0100    INFO    fil-consensus   filcns/upgrades.go:2418 All workers done after 4m34.3291245s
2024-01-25T16:12:03.287+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Result writer wrote 2947603 results to state tree after 4m34.329144208s
2024-01-25T16:12:03.287+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Running deferred migrations
2024-01-25T16:12:03.287+0100    INFO    fil-consensus   filcns/upgrades.go:2418 Running deferred migration for f05
2024-01-25T16:14:38.703+0100    INFO    fil-consensus   filcns/upgrades.go:2418 All 2947604 done after 7m9.743619s (6859/s)
2024-01-25T16:14:40.475+0100    ERROR   lotus-shed      lotus-shed/main.go:160  got different results with and without the cache: bafy2bzacecbl6cpzfynizqpoddnunz4pikd5ikdm7jxi3cipbm56jrkuio5sa, bafy2bzacedbolm3bvanxqy7b2n5xxovg2eel3ii35qlfunf2iktkteucd5nb2:
    main.glob..func97
        /Users/magik6k/lotus/cmd/lotus-shed/migrations.go:219

```

## Checklist

Before you mark the PR ready for review, please make sure that:

- [ ] Commits have a clear commit message.
- [ ] PR title is in the form of of `<PR type>: <area>: <change being made>`
  - example: ` fix: mempool: Introduce a cache for valid signatures`
  - `PR type`: fix, feat, build, chore, ci, docs, perf, refactor, revert, style, test
  - `area`, e.g. api, chain, state, market, mempool, multisig, networking, paych, proving, sealing, wallet, deps
- [ ] If the PR affects users (e.g., new feature, bug fix, system requirements change), update the CHANGELOG.md and add details to the UNRELEASED section.
- [ ] New features have usage guidelines and / or documentation updates in
  - [ ] [Lotus Documentation](https://lotus.filecoin.io)
  - [ ] [Discussion Tutorials](https://github.com/filecoin-project/lotus/discussions/categories/tutorials)
- [ ] Tests exist for new functionality or change in behavior
- [ ] CI is green